Backend/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pydantic==2.4.2
numpy==1.26.0
soundfile==0.12.1
python-dotenv==1.0.0
torch==2.1.0
transformers==4.35.0
pydub==0.25.1
aiohttp==3.8.5
openai==1.47.1
librosa==0.10.1
scipy==1.11.3